refactor: folder structure of scan checkin
