+++ lib/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_detail_clock.dart
@@ -170,7 +170,7 @@ class GuestEventDetailClock extends StatelessWidget {
                                   Text(
                                     durationToEvent != null
                                         ? t.event.eventStartIn(
-                                            time: printDuration(
+                                            time: prettyDuration(
                                               durationToEvent!,
                                               tersity: (durationToEvent
                                                               ?.inDays ??
