+++ lib/core/presentation/pages/event/event_detail_page/host_event_detail_page/widgets/host_event_basic_info_card.dart
@@ -136,7 +136,7 @@ class HostEventBasicInfoCard extends StatelessWidget {
                                   Text(
                                     durationToEvent != null
                                         ? t.event.eventStartIn(
-                                            time: printDuration(
+                                            time: prettyDuration(
                                               durationToEvent!,
                                               tersity: (durationToEvent
                                                               ?.inDays ??
