+++ lib/core/presentation/pages/scan_qr_code/views/scan_qr_rewards_view.dart
@@ -1,0 +1,1 @@
+
